Moved large chunks of code from main.py to main_utils.py
